requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
WTForms==3.1.1
email-validator==2.1.0
bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.0

SQLAlchemy>=2.0.30
Werkzeug==3.0.1
itsdangerous==2.1.2
cryptography==41.0.7
